+++ src/spyglass/common/common_usage.py
@@ -15,7 +15,7 @@ from pynwb import NWBHDF5IO
 
 from spyglass.common.common_nwbfile import AnalysisNwbfile, Nwbfile
 from spyglass.settings import export_dir, test_mode
-from spyglass.utils import SpyglassMixin, logger
+from spyglass.utils import SpyglassMixin, SpyglassMixinPart, logger
 from spyglass.utils.dj_graph import RestrGraph
 from spyglass.utils.dj_helper_fn import (
     make_file_obj_id_unique,
@@ -88,7 +88,7 @@ class ExportSelection(SpyglassMixin, dj.Manual):
     unique index (paper_id, analysis_id)
     """
 
-    class Table(SpyglassMixin, dj.Part):
+    class Table(SpyglassMixinPart):
         definition = """
         -> master
         table_id: int
@@ -100,6 +100,7 @@ class ExportSelection(SpyglassMixin, dj.Manual):
         def insert1(self, key, **kwargs):
             """Override insert1 to auto-increment table_id."""
             key = self._auto_increment(key, pk="table_id")
+            kwargs["skip_duplicates"] = True
             super().insert1(key, **kwargs)
 
         def insert(self, keys: List[dict], **kwargs):
@@ -107,6 +108,7 @@ class ExportSelection(SpyglassMixin, dj.Manual):
             if not isinstance(keys[0], dict):
                 raise TypeError("Pass Table Keys as list of dict")
             keys = [self._auto_increment(k, pk="table_id") for k in keys]
+            kwargs["skip_duplicates"] = True
             super().insert(keys, **kwargs)
 
     class File(SpyglassMixin, dj.Part):